Separate functional to files
