+++ CloudyCider/SceneDelegate.swift
@@ -23,8 +23,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     #else
     let window = UIWindow(frame: UIScreen.main.bounds)
     #endif
-    let ec2Store = EC2Store()
-    let controller = UIHostingController(rootView: HomeView().environmentObject(ec2Store))
+    let controller = UIHostingController(rootView: HomeView())
     window.rootViewController = controller
     self.window = window
     window.makeKeyAndVisible()
